fix(web): extract get chat id to hook
